app/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Stephen Young <stephen@theyoungs.cloud>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyexiftool = "^0.5.6"
pysiril = {file = "packages/pysiril-0.0.15-py3-none-any.whl"}
numpy = "^1.26.4"
matplotlib = "^3.8.3"
opencv-python = "^4.9.0.80"
astropy = "^6.1.1"
pygments = "^2.18.0"
redmail = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
